Revert to merge with trunk version 237367
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gupc@237999 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libcpp/line-map.c
@@ -102,7 +102,7 @@ location_adhoc_data_eq (const void *l1, const void *l2)
 static int
 location_adhoc_data_update (void **slot, void *data)
 {
-  *((char **) slot) += *((int64_t *) data);
+  *((char **) slot) += *((long long *) data);
   return 1;
 }
 
@@ -224,7 +224,7 @@ get_combined_adhoc_loc (struct line_maps *set,
 	  set->location_adhoc_data_map.allocated)
 	{
 	  char *orig_data = (char *) set->location_adhoc_data_map.data;
-	  int64_t offset;
+	  long long offset;
 	  /* Cast away extern "C" from the type of xrealloc.  */
 	  line_map_realloc reallocator = (set->reallocator
 					  ? set->reallocator
